+++ samples/src/main/java/me/ibrahimyilmaz/kiel/samples/adapter/viewholder/ImageMessageViewHolder.kt
@@ -1,12 +1,11 @@
-package me.ibrahimyilmaz.kiel.samples.datasource.viewbinder
+package me.ibrahimyilmaz.kiel.samples.adapter.viewholder
 
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
-import com.bumptech.glide.Glide
 import me.ibrahimyilmaz.kiel.adapter.RecyclerViewHolder
 import me.ibrahimyilmaz.kiel.samples.R
-import me.ibrahimyilmaz.kiel.samples.datasource.model.MessageViewState.Image
+import me.ibrahimyilmaz.kiel.samples.adapter.model.MessageViewState.Image
 
 class ImageMessageViewHolder(view: View) : RecyclerViewHolder<Image>(view) {
     val messageImage = view.findViewById<ImageView>(R.id.messageImage)