+++ samples/src/main/java/me/ibrahimyilmaz/kiel/samples/adapter/viewholder/PollMessageViewHolder.kt
@@ -1,11 +1,11 @@
-package me.ibrahimyilmaz.kiel.samples.datasource.viewbinder
+package me.ibrahimyilmaz.kiel.samples.adapter.viewholder
 
 import android.view.View
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import me.ibrahimyilmaz.kiel.adapter.RecyclerViewHolder
 import me.ibrahimyilmaz.kiel.samples.R
-import me.ibrahimyilmaz.kiel.samples.datasource.model.MessageViewState.Poll
+import me.ibrahimyilmaz.kiel.samples.adapter.model.MessageViewState.Poll
 
 class PollMessageViewHolder(view: View) : RecyclerViewHolder<Poll>(view) {
     val pollTitle = view.findViewById<TextView>(R.id.pollTitle)