+++ samples/src/main/java/me/ibrahimyilmaz/kiel/samples/adapter/viewholder/PollOptionViewHolder.kt
@@ -1,10 +1,10 @@
-package me.ibrahimyilmaz.kiel.samples.datasource.viewbinder
+package me.ibrahimyilmaz.kiel.samples.adapter.viewholder
 
 import android.view.View
 import android.widget.RadioButton
 import me.ibrahimyilmaz.kiel.adapter.RecyclerViewHolder
 import me.ibrahimyilmaz.kiel.samples.R
-import me.ibrahimyilmaz.kiel.samples.datasource.model.MessageViewState.Poll.PollOption
+import me.ibrahimyilmaz.kiel.samples.adapter.model.MessageViewState.Poll.PollOption
 
 
 class PollOptionViewHolder(view: View) : RecyclerViewHolder<PollOption>(view) {